Replant-World-divum-pwa_admin_changes/backend/pyproject.toml
[tool.poetry]
name = "replant"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.1"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
python-json-logger = "^2.0.7"
colorlog = "^6.8.0"
django-cors-headers = "^4.3.1"
drf-spectacular = "^0.27.0"
whitenoise = "^6.6.0"
django-debug-toolbar = "^4.2.0"
pycountry = "^23.12.11"
django-phonenumber-field = { extras = ["phonenumberslite"], version = "^7.3.0" }
django-filter = "^23.5"
cosmpy = "^0.9.2"
django-fsm = "^2.8.1"
pillow = "^10.2.0"
drf-extra-fields = "^3.7.0"
gunicorn = "^21.2.0"
django-storages = { extras = ["s3"], version = "^1.14.2" }
django-click = "^2.3.0"
django-admin-autocomplete-filter = "^0.7.1"
retry = "^0.9.2"
django-admin-rangefilter = "^0.12.1"
sendgrid = "^6.11.0"
django-mjml = "^1.2"
opensimplex = "^0.4.5"
scikit-learn = "^1.4.1.post1"
boto3-stubs = {extras = ["essential"], version = "^1.35.69"}
pydantic = "^2.10.1"
types-protobuf = "^5.28.3.20241030"
responses = "^0.25.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.6.0"
django-stubs = { extras = ["compatible-mypy"], version = "4.2.6" }
djangorestframework-stubs = { extras = [
    "compatible-mypy",
], version = "^3.14.4" }
django-types = "^0.19.1"
pytest = "^7.4.4"
pytest-django = "^4.7.0"
pytest-xdist = "^3.5.0"
model-bakery = "^1.17.0"
time-machine = "^2.13.0"
pytest-mock = "^3.12.0"
types-retry = "^0.9.9.4"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "replant.settings"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:The django.utils.baseconv module is deprecated",
]
markers = [
    "live_request: marks tests that call external services"
]

[tool.black]
exclude = "(.venv|migrations)"
target_version = ["py312"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
skip = [".venv", "migrations"]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = "migrations"
ignore_missing_imports = true
check_untyped_defs = true
show_column_numbers = true
warn_unused_ignores = true
strict_equality = true
no_implicit_optional = true
follow_imports = "silent"
mypy_path = "backend"

[tool.django-stubs]
django_settings_module = "replant.settings"
